requirements.txt
tensorflow>=2.12
numpy
matplotlib
Pillow
requests